Move attendance populating form to camera page
