+++ src/public/components/CameraPage.js
@@ -4,6 +4,10 @@ import keydown, { Keys } from 'react-keydown';
 import { queryGallery } from './requests/gallery';
 import Select from 'react-select';
 import Spinner from './Spinner';
+import DateTime from 'react-widgets/lib/DateTimePicker';
+import { getClasses } from './requests/classes';
+import Moment from 'moment';
+import { storeAttendanceRecord, emailLateStudents } from './requests/students';
 
 export default class CameraPage extends React.Component {
 
@@ -12,10 +16,23 @@ export default class CameraPage extends React.Component {
 
     this.state = {
       spinner: false,
-      checkedinUser: null
+      checkedinUser: null,
+      value: '',
+      selectedTimeCutoff: null,
+      options: [],
+      attendancePopulated: false,
+
     };
 
-    this.takeScreenshot = this.takeScreenshot.bind(this);
+    ['takeScreenshot',
+    'sendLateEmails',
+    'populateAttendanceRecord',
+    'updateSelectedTimeCutoff',
+    'getSelectOptions',
+    'handleSelectChange',
+    'toggleOff'].forEach((method) => {
+      this[method] = this[method].bind(this);
+    });
   }
 
   componentWillMount() {
@@ -34,20 +51,82 @@ export default class CameraPage extends React.Component {
     this.setState({ spinner: false, checkedinUser: 'hardcoded guy checked in' });
   }
 
+  async sendLateEmails () {
+    await emailLateStudents(this.state.selectedTimeCutoff);
+  }
+
+  async populateAttendanceRecord() {
+    if (this.state.value && this.state.selectedTimeCutoff) {
+      this.setState({ spinner: true, statusUpdated: false });
+      this.setState({ attendancePopulated: await storeAttendanceRecord(this.state.value, this.state.selectedTimeCutoff) });
+      this.setState({ spinner: false });
+    } else {
+      alert('Select Class(es) and Cutoff Time!');
+    }
+    this.toggleOff('attendancePopulated', 'value', 'selectedTimeCutoff');
+  }
+
+  updateSelectedTimeCutoff(e) {
+    let date = Moment(e);
+    this.setState({ selectedTimeCutoff: date });
+  }
+
+  async getSelectOptions() {
+    const classList = await getClasses();
+    const classes = classList.classes.map((classname) => {
+      return { label: classname, value: classname };
+    });
+    this.setState({ options: classes });
+  }
+
+  handleSelectChange(value) {
+    this.setState({ value });
+  }
+
+  toggleOff(status, ...states) {
+    setTimeout(() => {
+      this.setState({ [status]: false });
+      states.forEach((state) => {
+        this.setState({ [state]: false});
+      });
+    }, 5000);
+  }
+
   render() {
     return (
       <div>
 
         { this.state.mounted && <div><Webcam ref='webcam'/></div> }
 
-        <h1> Screenshots </h1>
-        
         <div>
           <button className="screenShotButton" onClick={this.takeScreenshot}>Take Screenshot</button>
         </div>
 
         {this.state.spinner && <Spinner/>}
         {this.state.checkedinUser}
+
+        <hr/>
+        <h3>Start Daily Attendance</h3>
+        Class:
+        <div className="classSelect" onClick={!this.state.options.length && this.getSelectOptions}>
+          <Select 
+            multi={true}
+            simpleValue
+            value={this.state.value}
+            placeholder="Select Class(es)..."
+            options={this.state.options}
+            onChange={this.handleSelectChange}
+          />
+        </div><br/>
+        Cutoff Time:
+        <DateTime
+          placeholder="Select Cutoff Time..."
+          onChange={this.updateSelectedTimeCutoff}
+          calendar={false}
+        /><br/>
+        <button className="populateAttendanceRecord" onClick={this.populateAttendanceRecord}>Populate Attendance Records</button><br/><br/>
+        {!this.state.attendancePopulated ? null : <h5>Populated daily attendance for {this.state.value} on {this.state.selectedTimeCutoff.format('dddd, MMM Do, YYYY')}!</h5>}
+        <button className="lateStudentButton" onClick={this.sendLateEmails}>Send Email to Late Students</button><hr/>
 
       </div>
     );
